1. Like, comment, subscribe  (automated command)

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,20 +1,20 @@
 #include <stdio.h>
 void main(void)
 {
-	long long int firstNum = 0;
+	unsigned long firstNum = 0;
 
-	long long int secondNum = 1;
+	unsigned long  secondNum = 1;
 
-	printf("%lld, %lld,  ", firstNum, secondNum);
+	printf("%lu, %lu,  ", firstNum, secondNum);
 	for (size_t i = 0; i < 50; i++)
 	{
-		long long int nextNumber = firstNum + secondNum;
+		unsigned long nextNumber = firstNum + secondNum;
 
 		firstNum = secondNum;
 
 		secondNum = nextNumber;
 
-		printf("%lld  ", secondNum);
+		printf("%lu  ", secondNum);
 		if (i != 49)
 		{
 			putchar(',');
